Fix Servlet instrumentation for servlet-api 3.1+
* Separate Servlet 3.0 and Servlet 3.1 instrumentations
* Separate Jetty 8.0 (depends on Servlet 3.0) and Jetty 9.1 (depends on Servlet 3.1) instrumentations
