+++ instrumentation/dropwizard-testing/dropwizard-testing.gradle
@@ -5,7 +5,7 @@ apply from: "$rootDir/gradle/instrumentation.gradle"
 
 dependencies {
   testImplementation project(':instrumentation:jaxrs:jaxrs-2.0')
-  testImplementation project(':instrumentation:servlet:servlet-3.0')
+  testImplementation project(':instrumentation:servlet:servlet-3:servlet-3.1')
 
   // First version with DropwizardTestSupport:
   testImplementation group: 'io.dropwizard', name: 'dropwizard-testing', version: '0.8.0'
