--- instrumentation/servlet/servlet-3.0/src/main/java/io/opentelemetry/instrumentation/auto/servlet/v3_0/AsyncContextInstrumentation.java
@@ -1,118 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.opentelemetry.instrumentation.auto.servlet.v3_0;
-
-import static io.opentelemetry.instrumentation.api.tracer.HttpServerTracer.CONTEXT_ATTRIBUTE;
-import static io.opentelemetry.javaagent.tooling.ClassLoaderMatcher.hasClassesNamed;
-import static io.opentelemetry.javaagent.tooling.bytebuddy.matcher.AgentElementMatchers.implementsInterface;
-import static io.opentelemetry.trace.TracingContextUtils.getSpan;
-import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.isPublic;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-
-import com.google.auto.service.AutoService;
-import io.grpc.Context;
-import io.opentelemetry.instrumentation.auto.api.CallDepthThreadLocalMap;
-import io.opentelemetry.javaagent.tooling.Instrumenter;
-import io.opentelemetry.trace.Span;
-import java.util.Map;
-import javax.servlet.AsyncContext;
-import javax.servlet.ServletRequest;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-
-@AutoService(Instrumenter.class)
-public final class AsyncContextInstrumentation extends Instrumenter.Default {
-
-  public AsyncContextInstrumentation() {
-    super("servlet", "servlet-3");
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      "io.opentelemetry.instrumentation.servlet.HttpServletRequestGetter",
-      "io.opentelemetry.instrumentation.servlet.ServletHttpServerTracer",
-      packageName + ".CountingHttpServletResponse",
-      packageName + ".CountingHttpServletResponse$CountingServletOutputStream",
-      packageName + ".CountingHttpServletResponse$CountingPrintWriter",
-      packageName + ".TagSettingAsyncListener",
-      packageName + ".Servlet3HttpServerTracer"
-    };
-  }
-
-  @Override
-  public ElementMatcher<ClassLoader> classLoaderMatcher() {
-    // Optimization for expensive typeMatcher.
-    return hasClassesNamed("javax.servlet.AsyncContext");
-  }
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return implementsInterface(named("javax.servlet.AsyncContext"));
-  }
-
-  @Override
-  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return singletonMap(
-        isMethod().and(isPublic()).and(named("dispatch")),
-        AsyncContextInstrumentation.class.getName() + "$DispatchAdvice");
-  }
-
-  /**
-   * When a request is dispatched, we want new request to have propagation headers from its parent
-   * request. The parent request's span is later closed by {@code
-   * TagSettingAsyncListener#onStartAsync}
-   */
-  public static class DispatchAdvice {
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static boolean enter(
-        @Advice.This AsyncContext context, @Advice.AllArguments Object[] args) {
-      int depth = CallDepthThreadLocalMap.incrementCallDepth(AsyncContext.class);
-      if (depth > 0) {
-        return false;
-      }
-
-      ServletRequest request = context.getRequest();
-
-      Context currentContext = Context.current();
-      Span currentSpan = getSpan(currentContext);
-      if (currentSpan.getContext().isValid()) {
-        // this tells the dispatched servlet to use the current span as the parent for its work
-        // (if the currentSpan is not valid for some reason, the original servlet span should still
-        // be present in the same request attribute, and so that will be used)
-        //
-        // the original servlet span stored in the same request attribute does not need to be saved
-        // and restored on method exit, because dispatch() hands off control of the request
-        // processing, and nothing can be done with the request anymore after this
-        request.setAttribute(CONTEXT_ATTRIBUTE, currentContext);
-      }
-
-      return true;
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void exit(@Advice.Enter boolean topLevel) {
-      if (topLevel) {
-        CallDepthThreadLocalMap.reset(AsyncContext.class);
-      }
-    }
-  }
-}
