--- instrumentation/servlet/servlet-3.0/src/test/resources/realm.properties
@@ -1,1 +1,0 @@
-user:password,role
