+++ instrumentation/servlet/servlet-3/servlet-3-testing/servlet-3-testing.gradle
@@ -1,0 +1,27 @@
+ext {
+  skipPublish = true
+}
+
+apply from: "$rootDir/gradle/instrumentation.gradle"
+
+dependencies {
+  testImplementation project(':instrumentation:servlet:servlet-3:servlet-3.0')
+  testImplementation project(':instrumentation:servlet:servlet-3:servlet-3.1')
+
+  testImplementation(project(':testing-common')) {
+    exclude group: 'org.eclipse.jetty', module: 'jetty-server'
+  }
+
+  testLibrary group: 'org.eclipse.jetty', name: 'jetty-server', version: '8.0.0.v20110901'
+  testLibrary group: 'org.eclipse.jetty', name: 'jetty-servlet', version: '8.0.0.v20110901'
+  testLibrary group: 'org.apache.tomcat.embed', name: 'tomcat-embed-core', version: '8.0.41'
+  testLibrary group: 'org.apache.tomcat.embed', name: 'tomcat-embed-jasper', version: '8.0.41'
+
+  // Jetty 10 seems to refuse to run on java8.
+  // TODO: we need to setup separate test for Jetty 10 when that is released.
+  latestDepTestLibrary group: 'org.eclipse.jetty', name: 'jetty-server', version: '9.+'
+  latestDepTestLibrary group: 'org.eclipse.jetty', name: 'jetty-servlet', version: '9.+'
+
+  latestDepTestLibrary group: 'org.apache.tomcat.embed', name: 'tomcat-embed-core', version: '9.+'
+  latestDepTestLibrary group: 'org.apache.tomcat.embed', name: 'tomcat-embed-jasper', version: '9.+'
+}
