+++ instrumentation/servlet/servlet-3/servlet-3.1/servlet-3.1.gradle
@@ -1,0 +1,16 @@
+apply from: "$rootDir/gradle/instrumentation.gradle"
+
+muzzle {
+  pass {
+    group = "javax.servlet"
+    module = 'javax.servlet-api'
+    versions = "[3.1,)"
+  }
+}
+
+dependencies {
+  implementation(project(':instrumentation-core:servlet'))
+  implementation(project(':instrumentation:servlet:servlet-3:servlet-3-common'))
+
+  compileOnly group: 'javax.servlet', name: 'javax.servlet-api', version: '3.1.0'
+}
