+++ settings.gradle
@@ -111,7 +111,9 @@ include ':instrumentation:jaxrs-client:jaxrs-client-2.0:jaxrs-client-2.0-resteas
 include ':instrumentation:jdbc'
 include ':instrumentation:jedis:jedis-1.4'
 include ':instrumentation:jedis:jedis-3.0'
-include ':instrumentation:jetty-8.0'
+include ':instrumentation:jetty:jetty-8.0'
+include ':instrumentation:jetty:jetty-9.1'
+include ':instrumentation:jetty:jetty-testing'
 include ':instrumentation:jms-1.1'
 include ':instrumentation:jsp-2.3'
 include ':instrumentation:kafka-clients-0.11'
@@ -150,7 +152,10 @@ include ':instrumentation:rxjava-1.0'
 include ':instrumentation:servlet:glassfish-testing'
 include ':instrumentation:servlet:servlet-common'
 include ':instrumentation:servlet:servlet-2.2'
-include ':instrumentation:servlet:servlet-3.0'
+include ':instrumentation:servlet:servlet-3:servlet-3-common'
+include ':instrumentation:servlet:servlet-3:servlet-3-testing'
+include ':instrumentation:servlet:servlet-3:servlet-3.0'
+include ':instrumentation:servlet:servlet-3:servlet-3.1'
 include ':instrumentation:spark-web-framework-2.3'
 include ':instrumentation:spring:spring-data-1.8'
 include ':instrumentation:spring:spring-scheduling-3.1'
